+++ packages/xchain-cosmos/src/util.ts
@@ -82,7 +82,12 @@ export const getTxsFromHistory = (txs: Array<TxResponse>, mainAsset: Asset): Txs
           .map((coin) => baseAmount(coin.amount, 6))
           .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), 6), baseAmount(0, 6))
 
-        const from_index = from.findIndex((value) => value.from === msgSend.from_address.toBech32())
+        let from_index = -1
+
+        from.forEach((value, index) => {
+          if (value.from === msgSend.from_address.toBech32()) from_index = index
+        })
+
         if (from_index === -1) {
           from.push({
             from: msgSend.from_address.toBech32(),
@@ -92,7 +97,12 @@ export const getTxsFromHistory = (txs: Array<TxResponse>, mainAsset: Asset): Txs
           from[from_index].amount = baseAmount(from[from_index].amount.amount().plus(amount.amount()), 6)
         }
 
-        const to_index = to.findIndex((value) => value.to === msgSend.to_address.toBech32())
+        let to_index = -1
+
+        to.forEach((value, index) => {
+          if (value.to === msgSend.to_address.toBech32()) to_index = index
+        })
+
         if (to_index === -1) {
           to.push({
             to: msgSend.to_address.toBech32(),
@@ -108,7 +118,13 @@ export const getTxsFromHistory = (txs: Array<TxResponse>, mainAsset: Asset): Txs
           const amount = input.coins
             .map((coin) => baseAmount(coin.amount, 6))
             .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), 6), baseAmount(0, 6))
-          const from_index = from.findIndex((value) => value.from === input.address)
+
+          let from_index = -1
+
+          from.forEach((value, index) => {
+            if (value.from === input.address) from_index = index
+          })
+
           if (from_index === -1) {
             from.push({
               from: input.address,
@@ -123,7 +139,13 @@ export const getTxsFromHistory = (txs: Array<TxResponse>, mainAsset: Asset): Txs
           const amount = output.coins
             .map((coin) => baseAmount(coin.amount, 6))
             .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), 6), baseAmount(0, 6))
-          const to_index = to.findIndex((value) => value.to === output.address)
+
+          let to_index = -1
+
+          to.forEach((value, index) => {
+            if (value.to === output.address) to_index = index
+          })
+
           if (to_index === -1) {
             to.push({
               to: output.address,