+++ packages/xchain-util/src/chain.test.ts
@@ -4,11 +4,13 @@ describe('chain', () => {
   it('checks type Chain', () => {
     expect(isChain('BNB')).toBeTruthy()
     expect(isChain('BTC')).toBeTruthy()
+    expect(isChain('BCH')).toBeTruthy()
     expect(isChain('ETH')).toBeTruthy()
     expect(isChain('THOR')).toBeTruthy()
     expect(isChain('GAIA')).toBeTruthy()
     expect(isChain('POLKA')).toBeTruthy()
     expect(isChain('SOL')).toBeTruthy()
+    expect(isChain('LTC')).toBeTruthy()
     expect(isChain('')).toBeFalsy()
     expect(isChain('invalid')).toBeFalsy()
   })
@@ -18,6 +20,9 @@ describe('chain', () => {
     })
     it('returns string for BTC', () => {
       expect(chainToString('BTC')).toEqual('Bitcoin')
+    })
+    it('returns string for BCH', () => {
+      expect(chainToString('BCH')).toEqual('Bitcoin Cash')
     })
     it('returns string for ETH', () => {
       expect(chainToString('ETH')).toEqual('Ethereum')
@@ -33,6 +38,8 @@ describe('chain', () => {
     })
     it('returns string for SOL', () => {
       expect(chainToString('SOL')).toEqual('Solana')
+    it('returns string for LTC', () => {
+      expect(chainToString('LTC')).toEqual('Litecoin')
     })
   })
 })